Completed basic functionality (add on etrack at a time.

+++ getartistexcludes3.h
@@ -1,4 +1,4 @@
 #ifndef GETARTISTEXCLUDES3_H
 #define GETARTISTEXCLUDES3_H
-void getExtendedExcludes(long *_shistCount, unsigned long *_splaylistSize);
+void getExtendedExcludes(long *_shistCount, int *_splaylistSize);
 #endif // GETARTISTEXCLUDES3_H
